Implement the Window internal line counter This is an odd implementation detail in the GPU's Window feature.
It is the final thing required for us to get a perfect render
of DMG-ACID2!

DMG-ACID2 is not a comprehensive test of the
GPU, but that's still pretty big!

I have noticed the Window at the top of Donkey Kong is one line out
after finishing a level, so it looks like we still have at least one
GPU fix to do.